@@ -1,5 +1,5 @@
 {
-    "name": "Ping Logger",
+    "name": "ezoterik/ping-logger",
     "description": "Ping Logger Application",
     "keywords": [
         "ping",
@@ -11,7 +11,7 @@
     ],
     "homepage": "https://bitbucket.org/ezoterik/ping_logger",
     "type": "project",
-    "license": "GPL-2.0",
+    "license": "MIT",
     "authors": [
         {
             "name": "Oleg Balykin",
@@ -20,38 +20,46 @@
             "role": "Developer"
         }
     ],
-    "support": {
-        "source": "https://bitbucket.org/ezoterik/ping_logger"
-    },
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.5.0",
-        "yiisoft/yii2": "2.0.7",
-        "yiisoft/yii2-bootstrap": "*",
-        "yiisoft/yii2-swiftmailer": "*",
-        "kartik-v/yii2-icons": "1.4.*",
+        "php": ">=7.4.0",
+        "yiisoft/yii2": "2.0.37.*",
+        "yiisoft/yii2-bootstrap": "~2.0.0",
+        "yiisoft/yii2-swiftmailer": "~2.1.0",
+        "kartik-v/yii2-icons": "~1.4.5",
         "omnilight/yii2-momentjs": "1.0.*",
-        "baibaratsky/yii2-rollbar": "1.3.*",
-        "ezoterik/yii2-rollbar-js": "1.0.*",
+        "baibaratsky/yii2-rollbar": "~1.8.0",
         "bower-asset/react": "0.13.*"
     },
     "require-dev": {
-        "yiisoft/yii2-codeception": "*",
-        "yiisoft/yii2-debug": "*",
-        "yiisoft/yii2-gii": "*",
-        "yiisoft/yii2-faker": "*"
-    },
-    "suggest": {
-        "codeception/codeception": "Codeception, 1.8.*@dev is currently works well with Yii.",
-        "codeception/specify": "BDD style code blocks for PHPUnit and Codeception",
-        "codeception/verify": "BDD Assertions for PHPUnit and Codeception"
+        "yiisoft/yii2-debug": "~2.1.0",
+        "yiisoft/yii2-gii": "~2.1.0",
+        "yiisoft/yii2-faker": "~2.0.0",
+        "codeception/codeception": "^4.0",
+        "codeception/verify": "~0.5.0 || ~1.1.0",
+        "codeception/specify": "~0.4.6",
+        "symfony/browser-kit": ">=2.7 <=4.2.4",
+        "codeception/module-filesystem": "^1.0.0",
+        "codeception/module-yii2": "^1.0.0",
+        "codeception/module-asserts": "^1.0.0",
+        "yiisoft/yii2-coding-standards": "~2.0.4"
     },
     "config": {
-        "process-timeout": 1800
+        "preferred-install": {
+            "*": "dist"
+        },
+        "process-timeout": 1800,
+        "fxp-asset": {
+            "enabled": false
+        }
     },
     "scripts": {
+        "post-install-cmd": [
+            "yii\\composer\\Installer::postInstall"
+        ],
         "post-create-project-cmd": [
-            "yii\\composer\\Installer::postCreateProject"
+            "yii\\composer\\Installer::postCreateProject",
+            "yii\\composer\\Installer::postInstall"
         ]
     },
     "extra": {
@@ -62,14 +70,18 @@
                     "web/assets": "0777",
                     "yii": "0755"
                 }
-            ],
+            ]
+        },
+        "yii\\composer\\Installer::postInstall": {
             "generateCookieValidationKey": [
                 "config/web-local.php"
             ]
-        },
-        "asset-installer-paths": {
-            "npm-asset-library": "vendor/npm",
-            "bower-asset-library": "vendor/bower"
         }
-    }
+    },
+    "repositories": [
+        {
+            "type": "composer",
+            "url": "https://asset-packagist.org"
+        }
+    ]
 }
