@@ -7,6 +7,8 @@
 
 namespace app\widgets;
 
+use yii\bootstrap\Widget;
+
 /**
  * Alert widget renders a message from session flash. All flash messages are displayed
  * in the sequence they were assigned using setFlash. You can set message as following:
@@ -26,7 +28,7 @@ namespace app\widgets;
  * @author Kartik Visweswaran <kartikv2@gmail.com>
  * @author Alexander Makarov <sam@rmcreative.ru>
  */
-class Alert extends \yii\bootstrap\Widget
+class Alert extends Widget
 {
     /**
      * @var array the alert types configuration for the flash messages.
@@ -34,18 +36,18 @@ class Alert extends \yii\bootstrap\Widget
      * - $key is the name of the session flash variable
      * - $value is the bootstrap alert type (i.e. danger, success, info, warning)
      */
-    public $alertTypes = [
-        'error'   => 'alert-danger',
-        'danger'  => 'alert-danger',
+    public array $alertTypes = [
+        'error' => 'alert-danger',
+        'danger' => 'alert-danger',
         'success' => 'alert-success',
-        'info'    => 'alert-info',
-        'warning' => 'alert-warning'
+        'info' => 'alert-info',
+        'warning' => 'alert-warning',
     ];
 
     /**
      * @var array the options for rendering the close button tag.
      */
-    public $closeButton = [];
+    public array $closeButton = [];
 
     public function init()
     {
@@ -57,7 +59,7 @@ class Alert extends \yii\bootstrap\Widget
 
         foreach ($flashes as $type => $data) {
             if (isset($this->alertTypes[$type])) {
-                $data = (array) $data;
+                $data = (array)$data;
                 foreach ($data as $message) {
                     /* initialize css class for each alert box */
                     $this->options['class'] = $this->alertTypes[$type] . $appendCss;
