@@ -1,28 +1,15 @@
-<?php
-/**
- * @link http://www.yiiframework.com/
- * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
- */
+<?php /** @noinspection PhpMissingFieldTypeInspection */
 
 namespace app\assets;
 
 use yii\web\AssetBundle;
 
-/**
- * @author Qiang Xue <qiang.xue@gmail.com>
- * @since 2.0
- */
 class AppAsset extends AssetBundle
 {
     public $sourcePath = '@app/assets/app';
 
     public $css = [
         'css/site.css',
-    ];
-
-    public $js = [
-        'js/all.js',
     ];
 
     public $depends = [
