@@ -1,5 +1,11 @@
 <?php
 
+use yii\caching\FileCache;
+use yii\console\controllers\MigrateController;
+use yii\gii\Module as YiiGiiModule;
+use yii\log\FileTarget;
+use yii\mutex\FileMutex;
+
 $params = require __DIR__ . '/params-local.php';
 $db = require __DIR__ . '/db-local.php';
 
@@ -8,22 +14,30 @@ $config = [
     'basePath' => dirname(__DIR__),
     'bootstrap' => ['log'],
     'controllerNamespace' => 'app\commands',
-    'modules' => [
+    'modules' => [],
+    'controllerMap' => [
+        'migrate' => [
+            'class' => MigrateController::class,
+            'migrationPath' => null,
+            'migrationNamespaces' => [
+                'app\migrations',
+            ],
+        ],
     ],
     'components' => [
         'cache' => [
-            'class' => 'yii\caching\FileCache',
+            'class' => FileCache::class,
         ],
         'log' => [
             'targets' => [
                 [
-                    'class' => 'yii\log\FileTarget',
+                    'class' => FileTarget::class,
                     'levels' => ['error', 'warning'],
                 ],
             ],
         ],
         'mutex' => [
-            'class' => 'yii\mutex\FileMutex',
+            'class' => FileMutex::class,
         ],
         'db' => $db,
         'formatter' => [
@@ -43,7 +57,7 @@ if (YII_ENV_DEV) {
     // configuration adjustments for 'dev' environment
     $config['bootstrap'][] = 'gii';
     $config['modules']['gii'] = [
-        'class' => 'yii\gii\Module',
+        'class' => YiiGiiModule::class,
     ];
 }
 
