@@ -2,26 +2,21 @@
 
 namespace app\controllers;
 
-use app\models\Log;
-use app\models\Object;
-use Yii;
 use app\models\Group;
 use app\models\search\GroupSearch;
+use Yii;
 use yii\filters\AccessControl;
+use yii\filters\VerbFilter;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
-use yii\filters\VerbFilter;
 
-/**
- * GroupController implements the CRUD actions for Group model.
- */
 class GroupController extends Controller
 {
-    public function behaviors()
+    public function behaviors(): array
     {
         return [
             'access' => [
-                'class' => AccessControl::className(),
+                'class' => AccessControl::class,
                 'rules' => [
                     [
                         'actions' => ['index', 'view', 'create', 'update', 'delete'],
@@ -31,7 +26,7 @@ class GroupController extends Controller
                 ],
             ],
             'verbs' => [
-                'class' => VerbFilter::className(),
+                'class' => VerbFilter::class,
                 'actions' => [
                     'delete' => ['post'],
                 ],
@@ -39,13 +34,9 @@ class GroupController extends Controller
         ];
     }
 
-    /**
-     * Lists all Group models.
-     * @return mixed
-     */
     public function actionIndex()
     {
-        $searchModel = new GroupSearch;
+        $searchModel = new GroupSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
 
         return $this->render('index', [
@@ -54,48 +45,32 @@ class GroupController extends Controller
         ]);
     }
 
-    /**
-     * Displays a single Group model.
-     * @param integer $id
-     * @return mixed
-     */
-    public function actionView($id)
+    public function actionView(int $id)
     {
         $group = $this->findModel($id);
 
         return $this->render('view', [
             'model' => $group,
-            'objects' => $group->objects
+            'objects' => $group->objects,
         ]);
     }
 
-    /**
-     * Creates a new Group model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return mixed
-     */
     public function actionCreate()
     {
-        $model = new Group;
+        $model = new Group();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             Yii::$app->getSession()->setFlash('success', 'Группа успешно создана');
 
             return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
         }
+
+        return $this->render('create', [
+            'model' => $model,
+        ]);
     }
 
-    /**
-     * Updates an existing Group model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     */
-    public function actionUpdate($id)
+    public function actionUpdate(int $id)
     {
         $model = $this->findModel($id);
 
@@ -103,24 +78,17 @@ class GroupController extends Controller
             Yii::$app->getSession()->setFlash('success', 'Группа успешно изменена');
 
             return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
         }
+
+        return $this->render('update', [
+            'model' => $model,
+        ]);
     }
 
-    /**
-     * Deletes an existing Group model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     */
-    public function actionDelete($id)
+    public function actionDelete(int $id)
     {
         $group = $this->findModel($id);
 
-        //Удаляем группу
         $group->delete();
 
         Yii::$app->getSession()->setFlash('success', 'Группа успешно удалена');
@@ -128,19 +96,12 @@ class GroupController extends Controller
         return $this->redirect(['index']);
     }
 
-    /**
-     * Finds the Group model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return Group the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
+    protected function findModel(int $id): Group
     {
         if (($model = Group::findOne($id)) !== null) {
             return $model;
-        } else {
-            throw new NotFoundHttpException('The requested page does not exist.');
         }
+
+        throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
