@@ -3,24 +3,21 @@
 namespace app\controllers;
 
 use app\models\Log;
+use app\models\PingObject;
+use app\models\search\PingObjectSearch;
 use Yii;
-use app\models\Object;
-use app\models\search\ObjectSearch;
 use yii\filters\AccessControl;
+use yii\filters\VerbFilter;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
-use yii\filters\VerbFilter;
 
-/**
- * ObjectController implements the CRUD actions for Object model.
- */
 class ObjectController extends Controller
 {
-    public function behaviors()
+    public function behaviors(): array
     {
         return [
             'access' => [
-                'class' => AccessControl::className(),
+                'class' => AccessControl::class,
                 'rules' => [
                     [
                         'actions' => ['index', 'view', 'create', 'update', 'delete', 'delete-logs'],
@@ -30,7 +27,7 @@ class ObjectController extends Controller
                 ],
             ],
             'verbs' => [
-                'class' => VerbFilter::className(),
+                'class' => VerbFilter::class,
                 'actions' => [
                     'delete' => ['post'],
                     'delete-logs' => ['post'],
@@ -39,13 +36,9 @@ class ObjectController extends Controller
         ];
     }
 
-    /**
-     * Lists all Object models.
-     * @return mixed
-     */
     public function actionIndex()
     {
-        $searchModel = new ObjectSearch;
+        $searchModel = new PingObjectSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
 
         return $this->render('index', [
@@ -54,48 +47,32 @@ class ObjectController extends Controller
         ]);
     }
 
-    /**
-     * Displays a single Object model.
-     * @param string $id
-     * @return mixed
-     */
-    public function actionView($id)
+    public function actionView(int $id)
     {
         $object = $this->findModel($id);
 
         return $this->render('view', [
             'model' => $object,
-            'logs' => $object->getLogs()->limit(100)->all()
+            'logs' => $object->getLogs()->limit(100)->all(),
         ]);
     }
 
-    /**
-     * Creates a new Object model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return mixed
-     */
     public function actionCreate()
     {
-        $model = new Object;
+        $model = new PingObject();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             Yii::$app->getSession()->setFlash('success', 'Объект успешно создан');
 
             return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
         }
+
+        return $this->render('create', [
+            'model' => $model,
+        ]);
     }
 
-    /**
-     * Updates an existing Object model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param string $id
-     * @return mixed
-     */
-    public function actionUpdate($id)
+    public function actionUpdate(int $id)
     {
         $model = $this->findModel($id);
 
@@ -103,24 +80,17 @@ class ObjectController extends Controller
             Yii::$app->getSession()->setFlash('success', 'Объект успешно изменен');
 
             return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
         }
+
+        return $this->render('update', [
+            'model' => $model,
+        ]);
     }
 
-    /**
-     * Deletes an existing Object model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param string $id
-     * @return mixed
-     */
-    public function actionDelete($id)
+    public function actionDelete(int $id)
     {
         $object = $this->findModel($id);
 
-        //Удаляем объект
         $object->delete();
 
         Yii::$app->getSession()->setFlash('success', 'Объект успешно удален');
@@ -128,37 +98,23 @@ class ObjectController extends Controller
         return $this->redirect(['index']);
     }
 
-    /**
-     * Удаляет связанные логи
-     *
-     * @param int $id
-     * @return mixed
-     */
-    public function actionDeleteLogs($id)
+    public function actionDeleteLogs(int $id)
     {
         $object = $this->findModel($id);
 
-        //Удаляем связанные логи
         Log::deleteAll(['object_id' => $object->id]);
 
-        Yii::$app->getSession()->setFlash('success', 'История успешно очищенна');
+        Yii::$app->getSession()->setFlash('success', 'История успешно очищена');
 
         return $this->redirect(['view', 'id' => $object->id]);
     }
 
-    /**
-     * Finds the Object model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param string $id
-     * @return \app\models\Object the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
+    protected function findModel($id): PingObject
     {
-        if (($model = Object::findOne($id)) !== null) {
+        if (($model = PingObject::findOne($id)) !== null) {
             return $model;
-        } else {
-            throw new NotFoundHttpException('The requested page does not exist.');
         }
+
+        throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
