@@ -1,17 +1,11 @@
 <?php
 
-use yii\base\Event;
-use yii\db\Connection;
+use yii\db\Connection as YiiDbConnection;
 
 return [
-    'class' => 'yii\db\Connection',
+    'class' => YiiDbConnection::class,
     'dsn' => 'mysql:host=localhost;dbname=app',
     'username' => 'root',
     'password' => '',
     'charset' => 'utf8',
-    'on afterOpen' => function (Event $event) {
-        /** @var Connection $sender */
-        $sender = $event->sender;
-        $sender->createCommand('SET time_zone = "+00:00"')->execute();
-    },
 ];
