@@ -3,12 +3,12 @@
 return [
     'adminEmail' => 'admin@example.com',
     'users' => [
-        '100' => [
-            'id' => '100',
-            'username' => 'admin',
-            'password' => 'admin',
-            'authKey' => 'qkUj4F3cmmHUTPZeTEotEAwVNyYdToK',
-            'accessToken' => 'vXisUofvoi3HEcjFnfqukxC3CsXRnF',
-        ],
+        //'100' => [
+        //    'id' => '100',
+        //    'username' => 'admin',
+        //    'password' => 'admin',
+        //    'authKey' => 'qkUj4F3cmmHUTPZeTEotEAwVNyYdToK',
+        //    'accessToken' => 'vXisUofvoi3HEcjFnfqukxC3CsXRnF',
+        //],
     ],
 ];
