@@ -2,7 +2,7 @@
 /**
  * Message translations.
  *
- * This file is automatically generated by 'yii message' command.
+ * This file is automatically generated by 'yii message/extract' command.
  * It contains the localizable messages extracted from source code.
  * You may modify this file by translating the extracted messages.
  *
@@ -17,34 +17,48 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
-    'Incorrect username or password.' => 'Неправильное имя пользователя или пароль.',
-    'Create' => 'Создать',
+    'Address' => 'Адрес',
+    'Are you sure you want to clear the history of this object?' => 'Вы уверены, что хотите очистить историю этого объекта?',
+    'Are you sure you want to delete this item with all objects?' => 'Вы уверены, что хотите удалить этот элемент вместе со всеми объектами?',
+    'Are you sure you want to delete this item?' => 'Вы уверены, что хотите удалить этот элемент?',
+    'Average RTT' => 'Средний RTT',
+    'Clear history' => 'Очистить историю',
     'Create a group' => 'Создать группу',
     'Create a object' => 'Создать объект',
-    'Date' => 'Дата',
+    'Create' => 'Создать',
+    'Creating a group' => 'Создание группы',
+    'Creating an object' => 'Создание объекта',
     'Date of activity' => 'Дата активности',
+    'Date' => 'Дата',
     'Delete' => 'Удалить',
+    'Editing a group' => 'Редактирование группы',
+    'Editing an object' => 'Редактирование объекта',
+    'Editing' => 'Редактирование',
     'Event' => 'Событие',
+    'Events' => 'События',
     'Group' => 'Группа',
     'Groups' => 'Группы',
     'Home' => 'Домой',
     'ID' => 'ID',
+    'Incorrect username or password.' => 'Неправильное имя пользователя или пароль.',
+    'Information' => 'Информация',
     'IP' => 'IP',
+    'Is disable' => 'Отключен',
+    'Lock date' => 'Дата блокировки',
     'Login' => 'Войти',
     'Logout' => 'Выйти',
+    'Note' => 'Заметка',
     'Object' => 'Объект',
     'Objects' => 'Объекты',
     'Password' => 'Пароль',
     'Please fill out the following fields to login' => 'Пожалуйста, заполните следующие поля для входа',
-    'Port' => 'Порт',
     'Port TCP' => 'Порт TCP',
     'Port UDP' => 'Порт UDP',
     'Remember Me' => 'Запомнить меня',
     'Save' => 'Сохранить',
     'Status' => 'Статус',
     'Title' => 'Название',
+    'Update' => 'Редактировать',
     'Username' => 'Логин',
-    'Is Disable' => 'Отключен',
-    'Address' => 'Адрес',
-    'Note' => 'Заметка',
+    'View' => 'Просмотр',
 ];
