@@ -7,17 +7,40 @@
  *
  * In order to run this script from the web, you should copy it to the web root.
  * If you are using Linux you can create a hard link instead, using the following command:
- * ln requirements.php ../requirements.php
+ * ln ../requirements.php requirements.php
  */
 
 // you may need to adjust this path to the correct Yii framework path
-$frameworkPath = dirname(__FILE__) . '/vendor/yiisoft/yii2';
+// uncomment and adjust the following line if Yii is not located at the default path
+//$frameworkPath = dirname(__FILE__) . '/vendor/yiisoft/yii2';
 
-if (!is_dir($frameworkPath)) {
-    echo '<h1>Error</h1>';
-    echo '<p><strong>The path to yii framework seems to be incorrect.</strong></p>';
-    echo '<p>You need to install Yii framework via composer or adjust the framework path in file <abbr title="' . __FILE__ . '">' . basename(__FILE__) . '</abbr>.</p>';
-    echo '<p>Please refer to the <abbr title="' . dirname(__FILE__) . '/README.md">README</abbr> on how to install Yii.</p>';
+
+if (!isset($frameworkPath)) {
+    $searchPaths = [
+        dirname(__FILE__) . '/vendor/yiisoft/yii2',
+        dirname(__FILE__) . '/../vendor/yiisoft/yii2',
+    ];
+    foreach ($searchPaths as $path) {
+        if (is_dir($path)) {
+            $frameworkPath = $path;
+            break;
+        }
+    }
+}
+
+if (!isset($frameworkPath) || !is_dir($frameworkPath)) {
+    $message = "<h1>Error</h1>\n\n"
+        . "<p><strong>The path to yii framework seems to be incorrect.</strong></p>\n"
+        . '<p>You need to install Yii framework via composer or adjust the framework path in file <abbr title="' . __FILE__ . '">' . basename(__FILE__) . "</abbr>.</p>\n"
+        . '<p>Please refer to the <abbr title="' . dirname(__FILE__) . "/README.md\">README</abbr> on how to install Yii.</p>\n";
+
+    if (!empty($_SERVER['argv'])) {
+        // do not print HTML when used in console mode
+        echo strip_tags($message);
+    } else {
+        echo $message;
+    }
+    exit(1);
 }
 
 require_once($frameworkPath . '/requirements/YiiRequirementChecker.php');
@@ -53,21 +76,21 @@ $requirements = [
     'phpSafeMode' => [
         'name' => 'PHP safe mode',
         'mandatory' => false,
-        'condition' => $requirementsChecker->checkPhpIniOff("safe_mode"),
+        'condition' => $requirementsChecker->checkPhpIniOff('safe_mode'),
         'by' => 'File uploading and console command execution',
         'memo' => '"safe_mode" should be disabled at php.ini',
     ],
     'phpExposePhp' => [
         'name' => 'Expose PHP',
         'mandatory' => false,
-        'condition' => $requirementsChecker->checkPhpIniOff("expose_php"),
+        'condition' => $requirementsChecker->checkPhpIniOff('expose_php'),
         'by' => 'Security reasons',
         'memo' => '"expose_php" should be disabled at php.ini',
     ],
     'phpAllowUrlInclude' => [
         'name' => 'PHP allow url include',
         'mandatory' => false,
-        'condition' => $requirementsChecker->checkPhpIniOff("allow_url_include"),
+        'condition' => $requirementsChecker->checkPhpIniOff('allow_url_include'),
         'by' => 'Security reasons',
         'memo' => '"allow_url_include" should be disabled at php.ini',
     ],
@@ -79,4 +102,7 @@ $requirements = [
         'memo' => 'PHP mail SMTP server required',
     ],
 ];
-$requirementsChecker->checkYii()->check($requirements)->render();
+
+$result = $requirementsChecker->checkYii()->check($requirements)->getResult();
+$requirementsChecker->render();
+exit($result['summary']['errors'] === 0 ? 0 : 1);
