@@ -1,66 +1,92 @@
 <?php
-
-use yii\helpers\Html;
-use yii\grid\GridView;
-
 /**
- * @var yii\web\View $this
- * @var yii\data\ActiveDataProvider $dataProvider
- * @var app\models\search\GroupSearch $searchModel
+ * @var View $this
+ * @var ActiveDataProvider $dataProvider
+ * @var GroupSearch $searchModel
  */
 
+use app\models\Group;
+use app\models\search\GroupSearch;
+use kartik\icons\Icon;
+use yii\data\ActiveDataProvider;
+use yii\grid\ActionColumn;
+use yii\grid\GridView;
+use yii\helpers\Html;
+use yii\web\View;
+
 $this->title = Yii::t('app', 'Groups');
+
 $this->params['breadcrumbs'][] = $this->title;
-?>
-<div class="group-index">
 
-    <h1><?= Html::encode($this->title) ?></h1>
+echo Html::tag('h1', Html::encode($this->title));
 
-    <?php // echo $this->render('_search', ['model' => $searchModel]); ?>
+echo Html::tag('p', Html::a(Icon::show('plus') . Yii::t('app', 'Create a group'), ['create'], ['class' => 'btn btn-success']));
 
-    <p>
-        <?= Html::a(Yii::t('app', 'Create a group'), ['create'], ['class' => 'btn btn-success']) ?>
-    </p>
+echo GridView::widget([
+    'dataProvider' => $dataProvider,
+    'filterModel' => $searchModel,
+    'rowOptions' => function (Group $model) {
+        $res = [];
 
-    <?= GridView::widget([
-        'dataProvider' => $dataProvider,
-        'filterModel' => $searchModel,
-        'rowOptions' => function ($model, $key, $index, $grid) {
-            $res = [];
+        if ($model->is_disable) {
+            $res['class'] = 'warning';
+        }
 
-            if ($model->is_disable) {
-                $res['class'] = 'warning';
-            }
-
-            return $res;
-        },
-        'columns' => [
-            [
-                'attribute' => 'id',
-                'format' => 'text',
-                'headerOptions' => ['width' => 50]
-            ],
-            'name',
-            [
-                'attribute' => 'is_disable',
-                'format' => 'boolean',
-                'filter' => [0 => 'Нет', 1 => 'Да'],
-            ],
-            [
-                'class' => 'yii\grid\ActionColumn',
-                'headerOptions' => ['width' => 80],
-                'buttons' => [
-                    'delete' => function ($url, $model) {
-                        return Html::a('<span class="glyphicon glyphicon-trash"></span>', $url, [
-                            'title' => Yii::t('app', 'Delete'),
-                            'data-confirm' => 'Вы точно хотите удалить группу со всеми ее объектами?',
-                            'data-method' => 'post',
-                            'data-pjax' => '0',
-                        ]);
-                    }
-                ]
-            ],
+        return $res;
+    },
+    'columns' => [
+        [
+            'attribute' => 'id',
+            'headerOptions' => ['width' => 50],
         ],
-    ]); ?>
+        'name',
+        [
+            'attribute' => 'is_disable',
+            'format' => 'boolean',
+            'filter' => [0 => 'Нет', 1 => 'Да'],
+        ],
+        [
+            'class' => ActionColumn::class,
+            'contentOptions' => ['class' => 'action-column'],
+            'buttons' => [
+                'view' => function ($url) {
+                    $options = [
+                        'title' => Yii::t('app', 'View'),
+                        'data' => [
+                            'toggle' => 'tooltip',
+                            'pjax' => '0',
+                        ],
+                    ];
 
-</div>
+                    return Html::a(Icon::show('eye', ['framework' => Icon::FAR, 'space' => false]), $url, $options);
+                },
+                'update' => function ($url) {
+                    $options = [
+                        'title' => Yii::t('app', 'Update'),
+                        'data' => [
+                            'toggle' => 'tooltip',
+                            'pjax' => '0',
+                        ],
+                    ];
+
+                    return Html::a(Icon::show('edit', ['framework' => Icon::FAR, 'space' => false]), $url, $options);
+                },
+                'delete' => function ($url) {
+                    $options = [
+                        'title' => Yii::t('app', 'Delete'),
+                        'class' => 'delete',
+                        'data' => [
+                            'confirm' => Yii::t('app', 'Are you sure you want to delete this item with all objects?'),
+                            'method' => 'post',
+                            'toggle' => 'tooltip',
+                            'pjax' => '0',
+                        ],
+                    ];
+
+                    return Html::a(Icon::show('trash-alt', ['framework' => Icon::FAR, 'space' => false]), $url, $options);
+                },
+            ],
+            'headerOptions' => ['width' => 1],
+        ],
+    ],
+]);
