@@ -1,17 +1,19 @@
 <?php
-use app\models\User;
-use kartik\icons\Icon;
-use yii\helpers\Html;
-use yii\bootstrap\Nav;
-use yii\bootstrap\NavBar;
-use yii\widgets\Breadcrumbs;
-use app\assets\AppAsset;
-use app\widgets\Alert;
-
 /**
- * @var \yii\web\View $this
+ * @var View $this
  * @var string $content
  */
+
+use app\assets\AppAsset;
+use app\models\User;
+use app\widgets\Alert;
+use kartik\icons\Icon;
+use yii\bootstrap\Nav;
+use yii\bootstrap\NavBar;
+use yii\helpers\Html;
+use yii\web\View;
+use yii\widgets\Breadcrumbs;
+
 AppAsset::register($this);
 
 $this->registerJs("
@@ -20,22 +22,22 @@ $('[data-toggle=\"tooltip\"]').tooltip();
 ");
 
 Icon::map($this);
+
+$this->beginPage();
 ?>
-<?php $this->beginPage() ?>
 <!DOCTYPE html>
 <html lang="<?= Yii::$app->language ?>">
-<head>
-    <meta charset="<?= Yii::$app->charset ?>"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <?= Html::csrfMetaTags() ?>
-    <title><?= Html::encode($this->title) ?></title>
-    <?php $this->head() ?>
-</head>
-<body>
-
-<?php $this->beginBody() ?>
-    <div class="wrap">
-        <?php
+    <head>
+        <meta charset="<?= Yii::$app->charset ?>"/>
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <?= Html::csrfMetaTags() ?>
+        <title><?= Html::encode($this->title) ?></title>
+        <?php $this->head() ?>
+    </head>
+    <body>
+    <?php $this->beginBody() ?>
+        <div class="wrap">
+            <?php
             /** @var User $user */
             $user = Yii::$app->getUser()->getIdentity();
 
@@ -46,6 +48,7 @@ Icon::map($this);
                     'class' => 'navbar-inverse navbar-fixed-top',
                 ],
             ]);
+
             echo Nav::widget([
                 'options' => ['class' => 'navbar-nav navbar-right'],
                 'items' => [
@@ -69,31 +72,31 @@ Icon::map($this);
                         [
                             'label' => Icon::show('power-off') . Yii::t('app', 'Logout') . ' (' . Html::encode($user->username) . ')',
                             'url' => ['site/logout'],
-                            'linkOptions' => ['data-method' => 'post']
+                            'linkOptions' => ['data-method' => 'post'],
                         ],
                 ],
                 'encodeLabels' => false,
             ]);
             NavBar::end();
-        ?>
+            ?>
 
-        <div class="container">
-            <?= Breadcrumbs::widget([
-                'links' => isset($this->params['breadcrumbs']) ? $this->params['breadcrumbs'] : [],
-            ]) ?>
-            <?= Alert::widget() ?>
-            <?= $content ?>
+            <div class="container">
+                <?= Breadcrumbs::widget([
+                    'links' => isset($this->params['breadcrumbs']) ? $this->params['breadcrumbs'] : [],
+                ]) ?>
+                <?= Alert::widget() ?>
+                <?= $content ?>
+            </div>
         </div>
-    </div>
 
-    <footer class="footer">
-        <div class="container">
-            <p class="pull-left">&copy; Ping Logger <?= date('Y') ?></p>
-        </div>
-    </footer>
+        <footer class="footer">
+            <div class="container">
+                <p class="pull-left">&copy; Ping Logger <?= date('Y') ?></p>
+            </div>
+        </footer>
 
-<?php $this->endBody() ?>
-</body>
+        <?php $this->endBody() ?>
+    </body>
 </html>
 <?php
 $this->endPage();
