@@ -1,27 +1,19 @@
 <?php
+/**
+ * @var $this View
+ * @var $name string
+ * @var $message string
+ * @var $exception Exception
+ */
 
 use yii\helpers\Html;
-
-/* @var $this yii\web\View */
-/* @var $name string */
-/* @var $message string */
-/* @var $exception Exception */
+use yii\web\View;
 
 $this->title = $name;
 ?>
-<div class="site-error">
+<h1><?= Html::encode($this->title) ?></h1>
 
-    <h1><?= Html::encode($this->title) ?></h1>
+<div class="alert alert-danger"><?= nl2br(Html::encode($message)) ?></div>
 
-    <div class="alert alert-danger">
-        <?= nl2br(Html::encode($message)) ?>
-    </div>
-
-    <p>
-        The above error occurred while the Web server was processing your request.
-    </p>
-    <p>
-        Please contact us if you think this is a server error. Thank you.
-    </p>
-
-</div>
+<p>The above error occurred while the Web server was processing your request.</p>
+<p>Please contact us if you think this is a server error. Thank you.</p>
