@@ -1,41 +1,42 @@
 <?php
-use yii\helpers\Html;
-use yii\widgets\ActiveForm;
+/**
+ * @var $this View
+ * @var $form ActiveForm
+ * @var $model LoginForm
+ */
 
-/* @var $this yii\web\View */
-/* @var $form yii\widgets\ActiveForm */
-/* @var $model app\models\LoginForm */
+use app\models\form\LoginForm;
+use yii\bootstrap\ActiveForm;
+use yii\helpers\Html;
+use yii\web\View;
 
 $this->title = Yii::t('app', 'Login');
+
 $this->params['breadcrumbs'][] = $this->title;
+
+echo Html::tag('h1', Html::encode($this->title));
+
+echo Html::tag('p', Yii::t('app', 'Please fill out the following fields to login') . ':');
+
+$form = ActiveForm::begin([
+    'id' => 'login-form',
+    'options' => ['class' => 'form-horizontal'],
+    'fieldConfig' => [
+        'template' => "{label}\n<div class=\"col-lg-3\">{input}</div>\n<div class=\"col-lg-8\">{error}</div>",
+        'labelOptions' => ['class' => 'col-lg-1 control-label'],
+    ],
+]);
+
+echo $form->field($model, 'username');
+echo $form->field($model, 'password')->passwordInput();
+echo $form->field($model, 'rememberMe', [
+    'template' => "<div class=\"col-lg-offset-1 col-lg-3\">{input}</div>\n<div class=\"col-lg-8\">{error}</div>",
+])->checkbox();
 ?>
-<div class="site-login">
-    <h1><?= Html::encode($this->title) ?></h1>
-
-    <p><?= Yii::t('app', 'Please fill out the following fields to login') ?>:</p>
-
-    <?php $form = ActiveForm::begin([
-        'id' => 'login-form',
-        'options' => ['class' => 'form-horizontal'],
-        'fieldConfig' => [
-            'template' => "{label}\n<div class=\"col-lg-3\">{input}</div>\n<div class=\"col-lg-8\">{error}</div>",
-            'labelOptions' => ['class' => 'col-lg-1 control-label'],
-        ],
-    ]); ?>
-
-    <?= $form->field($model, 'username') ?>
-
-    <?= $form->field($model, 'password')->passwordInput() ?>
-
-    <?= $form->field($model, 'rememberMe', [
-        'template' => "<div class=\"col-lg-offset-1 col-lg-3\">{input}</div>\n<div class=\"col-lg-8\">{error}</div>",
-    ])->checkbox() ?>
-
     <div class="form-group">
         <div class="col-lg-offset-1 col-lg-11">
             <?= Html::submitButton(Yii::t('app', 'Login'), ['class' => 'btn btn-primary', 'name' => 'login-button']) ?>
         </div>
     </div>
-
-    <?php ActiveForm::end(); ?>
-</div>
+<?php
+ActiveForm::end();
